Combine VB comparers into one, and combine VB and C# comparers together (#51834)
