+++ src/EditorFeatures/VisualBasicTest/EditAndContinue/StatementEditingTests.vb
@@ -74,7 +74,7 @@ End If
             Dim m2 = MakeMethodBody(src2)
 
             Dim knownMatches = {New KeyValuePair(Of SyntaxNode, SyntaxNode)(m1, m2)}
-            Dim match = StatementSyntaxComparer.Default.ComputeMatch(m1, m2, knownMatches)
+            Dim match = SyntaxComparer.Statement.ComputeMatch(m1, m2, knownMatches)
             Dim actual = ToMatchingPairs(match)
 
             Dim expected = New MatchingPairs From
@@ -1334,7 +1334,7 @@ End Try
             Dim knownMatches = {New KeyValuePair(Of SyntaxNode, SyntaxNode)(m1.Statements(1), m2.Statements(0))}
 
             ' pre-matched:
-            Dim match = StatementSyntaxComparer.Default.ComputeMatch(m1, m2, knownMatches)
+            Dim match = SyntaxComparer.Statement.ComputeMatch(m1, m2, knownMatches)
             Dim actual = ToMatchingPairs(match)
 
             Dim expected = New MatchingPairs From
@@ -1348,7 +1348,7 @@ End Try
             expected.AssertEqual(actual)
 
             ' not pre-matched:
-            match = StatementSyntaxComparer.Default.ComputeMatch(m1, m2)
+            match = SyntaxComparer.Statement.ComputeMatch(m1, m2)
             actual = ToMatchingPairs(match)
 
             expected = New MatchingPairs From
